python/requirements.txt
# MarsFX Data Generator Requirements

# Core data processing (compatible with Python 3.13)
numpy>=1.26.0,<2.0.0
pandas>=2.2.0
pyarrow>=14.0.0

# Kafka streaming (compatible with Python 3.13)
kafka-python>=2.0.2
confluent-kafka>=2.6.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.1

# Time and date handling
python-dateutil==2.8.2

# Logging and monitoring
structlog==24.1.0

# Data validation (compatible with Python 3.13)
pydantic>=2.9.0
pydantic-settings>=2.5.0

# CLI
click==8.1.7
rich==13.7.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0

# Trino connectivity (for direct writes)
trino==0.328.0
sqlalchemy==2.0.25

# PostgreSQL (for metadata)
psycopg2-binary>=2.9.0
scipy>=1.11.0
